Remove references to the NewsLink model
Time to clean the code!

I keep the news_links table in database for now, I'll delete it in the
next version. It’s a protection against errors.
