+++ tests/bootstrap.php
@@ -235,29 +235,6 @@ echo 'Use SEED=' . $faker_seed . " to reproduce this suite.\n";
 );
 
 \Minz\Tests\DatabaseFactory::addFactory(
-    'news_link',
-    '\flusio\models\dao\NewsLink',
-    [
-        'created_at' => function () use ($faker) {
-            return $faker->iso8601;
-        },
-        'url' => function () use ($faker) {
-            return $faker->url;
-        },
-        'link_id' => function () {
-            $link_factory = new \Minz\Tests\DatabaseFactory('link');
-            return $link_factory->create([
-                'is_hidden' => 0,
-            ]);
-        },
-        'user_id' => function () use ($faker) {
-            $user_factory = new \Minz\Tests\DatabaseFactory('user');
-            return $user_factory->create();
-        },
-    ]
-);
-
-\Minz\Tests\DatabaseFactory::addFactory(
     'message',
     '\flusio\models\dao\Message',
     [
